changed promise location
